INT-2040 changed signature of MessageStore.getMessageCount() to return long. Changed affected implementations

+++ spring-integration-gemfire/src/main/java/org/springframework/integration/gemfire/store/GemfireMessageStore.java
@@ -50,7 +50,7 @@ public class GemfireMessageStore implements MessageStore {
 		return this.region.remove(id);
 	}
 
-	public int getMessageCount() {
+	public long getMessageCount() {
 		return this.region.size();
 	}
